feat: 스토리 모달 3차 UX
